% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_python_libraries.R
\name{check_python_libraries}
\alias{check_python_libraries}
\title{Check and Install Python Libraries (numpy and onnxruntime)}
\usage{
check_python_libraries()
}
\value{
A list indicating whether `numpy` and `onnxruntime` are installed.
The list contains the following logical elements:
\item{numpy_installed}{TRUE if `numpy` is installed, FALSE otherwise.}
\item{onnxruntime_installed}{TRUE if `onnxruntime` is installed, FALSE otherwise.}
}
\description{
This function checks whether the Python libraries `numpy` and `onnxruntime` are installed. If not, it will prompt
the user to decide whether to install them. If the user chooses 'y', the required library will be installed using
the `reticulate` package. If the user chooses 'n', the installation will be skipped. @seealso \link[EFAfactors]{DNN_predictor}
}
\examples{
# Check and install necessary Python libraries
check_python_libraries()

}
