% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_scaler}
\alias{load_scaler}
\title{Load the Scaler for the Pre-Trained Deep Neural Network (DNN)}
\usage{
load_scaler()
}
\value{
scaler objective.
}
\description{
Loads the scaler object within the \code{EFAfactors} package. This object is a \code{list} containing a mean vector and
a standard deviation vector, which were computed from the 10,000,000 datasets \link[EFAfactors]{data.datasets}
used for training the Pre-Trained Deep Neural Network (DNN). It serves as a tool for normalizing features in
\link[EFAfactors]{DNN_predictor}.
}
\examples{
library(EFAfactors)

scaler <- load_scaler()
print(scaler)


}
\seealso{
\link[EFAfactors]{DNN_predictor}, \link[EFAfactors]{normalizor}, \link[EFAfactors]{data.datasets}, \link[EFAfactors]{data.scaler}
}
