% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_xgb}
\alias{load_xgb}
\title{Load the Tuned XGBoost Model}
\usage{
load_xgb()
}
\value{
The tuned XGBoost model object
}
\description{
Loads the tuned XGBoost model object within the \code{EFAfactors} package
into the global environment and retrieves it for use. Only the core model is retained to reduce the size.
}
\examples{
library(EFAfactors)

xgb_model <- load_xgb()
print(xgb_model)


}
