% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.xgb.R
\docType{data}
\name{model.xgb}
\alias{model.xgb}
\title{the Tuned XGBoost Model for Determining the Number of Facotrs}
\format{
An object of class \code{TuneModel} is the Tuned XGBoost Model for Determining the Number of Facotrs
}
\description{
the Tuned XGBoost Model for Determining the Number of Facotrs
}
\examples{
data(model.xgb)
print(model.xgb)

model.xgb <- load_xgb()
print(model.xgb)

}
\seealso{
\link[EFAfactors]{FF}, \link[EFAfactors]{load_xgb}
}
