% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.DNN_predictor}
\alias{plot.DNN_predictor}
\title{Plot DNN Predictor Classification Probability Distribution}
\usage{
\method{plot}{DNN_predictor}(x, ...)
}
\arguments{
\item{x}{An object of class \code{DNN_predictor}, representing the results to be plotted.}

\item{...}{Additional arguments to be passed to the plotting function.}
}
\value{
None. This function is used for side effects (plotting).
}
\description{
This function generates a bar plot of the classification probabilities predicted by the pre-trained deep neural network
for determining the number of factors. The plot displays the probability distribution across different numbers of factors,
with each bar representing the probability for a specific number of factors. The maximum number of factors that the network
can evaluate is 10. The function also annotates each bar with its probability value.
}
\seealso{
\link[EFAfactors]{DNN_predictor}
}
