% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.PA}
\alias{plot.PA}
\title{Plot Parallel Analysis Scree Plot}
\usage{
\method{plot}{PA}(x, ...)
}
\arguments{
\item{x}{An object of class \code{PA}, representing the results to be plotted.}

\item{...}{Additional arguments to be passed to the plotting function.}
}
\value{
None. This function is used for side effects (plotting).
}
\description{
This function creates a Parallel Analysis (PA) scree plot to compare the eigenvalues of the actual
data with the eigenvalues from simulated data. The plot helps in determining the number of factors
by visualizing where the eigenvalues of the actual data intersect with those from simulated data.
It provides a graphical representation of the results from a parallel analysis to aid in factor selection.
}
\examples{
library(EFAfactors)
set.seed(123)

##Take the data.bfi dataset as an example.
data(data.bfi)

response <- as.matrix(data.bfi[, 1:25]) ## loading data
response <- na.omit(response) ## Remove samples with NA/missing values

## Transform the scores of reverse-scored items to normal scoring
response[, c(1, 9, 10, 11, 12, 22, 25)] <- 6 - response[, c(1, 9, 10, 11, 12, 22, 25)] + 1

\donttest{
 PA.obj <- PA(response)

 ## PA plot
 plot(PA.obj)

}


}
\seealso{
\link[EFAfactors]{PA}
}
