% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FF.R
\name{predictLearner.classif.xgboost.earlystop}
\alias{predictLearner.classif.xgboost.earlystop}
\title{Prediction Function for the Tuned XGBoost Model with Early Stopping}
\usage{
\method{predictLearner}{classif.xgboost.earlystop}(.learner, .model, .newdata, ...)
}
\arguments{
\item{.learner}{An object representing the learner.}

\item{.model}{The trained XGBoost model used to make predictions.}

\item{.newdata}{A data frame or matrix containing new observations for which predictions are to be made.}

\item{...}{Additional parameters passed to the \code{predict} function in XGBoost.}
}
\value{
A vector of predicted class labels or a matrix of predicted probabilities.
}
\description{
This function performs predictions using a trained XGBoost model with early stopping.
The function itself does not have any specific purpose; its existence is solely to
ensure the proper operation of \link[EFAfactors]{FF}.
}
