% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.CD}
\alias{print.CD}
\title{Print Comparison Data Method Results}
\usage{
\method{print}{CD}(x, ...)
}
\arguments{
\item{x}{An object of class \code{CD}, representing the results to be printed.}

\item{...}{Additional arguments to be passed to the plotting function.}
}
\value{
None. This function is used for side effects (printing).
}
\description{
This function prints the number of factors suggested by the Comparison Data (CD) method.
}
\seealso{
\link[EFAfactors]{CD}
}
