% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.CDF}
\alias{print.CDF}
\title{Print Comparison Data Forest (CDF) Results}
\usage{
\method{print}{CDF}(x, ...)
}
\arguments{
\item{x}{An object of class \code{CDF}, representing the results to be printed.}

\item{...}{Additional arguments to be passed to the plotting function.}
}
\value{
None. This function is used for side effects (printing).
}
\description{
This function prints the number of factors suggested by the Comparison Data Forest.
}
\seealso{
\link[EFAfactors]{CDF}
}
