% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.DNN_predictor}
\alias{print.DNN_predictor}
\title{Print DNN Predictor Method Results}
\usage{
\method{print}{DNN_predictor}(x, ...)
}
\arguments{
\item{x}{An object of class \code{DNN_predictor}, representing the results to be printed.}

\item{...}{Additional arguments to be passed to the plotting function.}
}
\value{
None. This function is used for side effects (printing).
}
\description{
This function prints the number of factors suggested by the deep neural network (DNN) predictor.
}
\seealso{
\link[EFAfactors]{DNN_predictor}
}
