% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.EFAdata}
\alias{print.EFAdata}
\title{Print the EFAsim.data}
\usage{
\method{print}{EFAdata}(x, ...)
}
\arguments{
\item{x}{An object of class \code{EFA.data}, representing the results to be printed.}

\item{...}{Additional arguments to be passed to the plotting function.}
}
\value{
None. This function is used for side effects (printing).
}
\description{
This function prints the items in factors.
}
\seealso{
\link[EFAfactors]{EFAsim.data}
}
