% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.EFAkmeans}
\alias{print.EFAkmeans}
\title{Print EFAkmeans Method Results}
\usage{
\method{print}{EFAkmeans}(x, ...)
}
\arguments{
\item{x}{An object of class \code{EFAkmeans}, representing the results to be printed.}

\item{...}{Additional arguments to be passed to the plotting function.}
}
\value{
None. This function is used for side effects (printing).
}
\description{
This function prints the number of factors suggested by the EFAkmeans method using the Second-Order Difference (SOD) approach.
}
\seealso{
\link[EFAfactors]{EFAkmeans}
}
