% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.EFAscreet}
\alias{print.EFAscreet}
\title{Print the Scree Plot}
\usage{
\method{print}{EFAscreet}(x, ...)
}
\arguments{
\item{x}{An object of class \code{EFAscreet}, which contains the eigenvalues from the factor analysis.}

\item{...}{Additional arguments to be passed to the plotting function.}
}
\value{
None. This function is used for side effects (printing).
}
\description{
Prints a brief summary of an object of class \code{EFAscreet}. This function will display the scree plot
of the eigenvalues when called, providing a visual representation of the factor analysis results.
}
