% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.EFAvote}
\alias{print.EFAvote}
\title{Print Voting Method Results}
\usage{
\method{print}{EFAvote}(x, ...)
}
\arguments{
\item{x}{An object of class \code{EFAvote}, representing the results to be printed.}

\item{...}{Additional arguments to be passed to the plotting function.}
}
\value{
None. This function is used for side effects (printing).
}
\description{
This function prints the number of factors suggested by the voting.
}
\seealso{
\link[EFAfactors]{EFAvote}
}
