% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.EKC}
\alias{print.EKC}
\title{Print Empirical Kaiser Criterion Results}
\usage{
\method{print}{EKC}(x, ...)
}
\arguments{
\item{x}{An object of class \code{EKC}, representing the results to be printed.}

\item{...}{Additional arguments to be passed to the plotting function.}
}
\value{
None. This function is used for side effects (printing).
}
\description{
This function prints the number of factors suggested by the Empirical Kaiser Criterion (EKC).
}
\seealso{
\link[EFAfactors]{EKC}
}
