% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.FF}
\alias{print.FF}
\title{Print Factor Forest (FF) Results}
\usage{
\method{print}{FF}(x, ...)
}
\arguments{
\item{x}{An object of class \code{FF}, representing the results to be printed.}

\item{...}{Additional arguments to be passed to the plotting function.}
}
\value{
None. This function is used for side effects (printing).
}
\description{
This function prints the number of factors suggested by the Factor Forest.
}
\seealso{
\link[EFAfactors]{FF}
}
