% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.Hull}
\alias{print.Hull}
\title{Print Hull Method Results}
\usage{
\method{print}{Hull}(x, ...)
}
\arguments{
\item{x}{An object of class \code{Hull}, representing the results to be printed.}

\item{...}{Additional arguments to be passed to the print function.}
}
\value{
None. This function is used for side effects (printing).
}
\description{
This function prints the number of factors suggested by the Hull method.
}
\seealso{
\link[EFAfactors]{Hull}
}
