% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.KGC}
\alias{print.KGC}
\title{Print Kaiser-Guttman Criterion Results}
\usage{
\method{print}{KGC}(x, ...)
}
\arguments{
\item{x}{An object of class \code{KGC}, representing the results to be printed.}

\item{...}{Additional arguments to be passed to the plotting function.}
}
\value{
None. This function is used for side effects (printing).
}
\description{
This function prints the number of factors suggested by the Kaiser-Guttman Criterion (KGC).
}
\seealso{
\link[EFAfactors]{KGC}
}
