% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.PA}
\alias{print.PA}
\title{Print Parallel Analysis Method Results}
\usage{
\method{print}{PA}(x, ...)
}
\arguments{
\item{x}{An object of class \code{PA}, representing the results to be printed.}

\item{...}{Additional arguments to be passed to the plotting function.}
}
\value{
None. This function is used for side effects (printing).
}
\description{
This function prints the number of factors suggested by the Parallel Analysis (PA) method.
}
\seealso{
\link[EFAfactors]{PA}
}
